move data typer
