+++ src/packages/property-editors/date-picker/manifests.ts
