+++ src/packages/property-editors/date-picker/property-editor-ui-date-picker.test.ts
@@ -1,5 +1,5 @@
 import { expect, fixture, html } from '@open-wc/testing';
-import type { UmbInputDateElement } from '../../../components/input-date/input-date.element.js';
+import type { UmbInputDateElement } from '../../core/components/input-date/input-date.element.js';
 import { UmbPropertyEditorUIDatePickerElement } from './property-editor-ui-date-picker.element.js';
 import { type UmbTestRunnerWindow, defaultA11yConfig } from '@umbraco-cms/internal/test-utils';
 import { UmbPropertyEditorConfigCollection } from '@umbraco-cms/backoffice/property-editor';
